[후원하기] main activity 로 이동 (팝업에서도 사용하는 문제 때문에...)
