+++ app/src/main/java/womenproject/com/mybury/presentation/base/BaseFragment.kt
@@ -8,6 +8,7 @@ import androidx.activity.OnBackPressedCallback
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
 import androidx.fragment.app.Fragment
+import womenproject.com.mybury.data.SupportInfo
 import womenproject.com.mybury.presentation.MainActivity
 
 
@@ -111,5 +112,22 @@ abstract class BaseFragment<T : ViewDataBinding, R : BaseViewModel> : Fragment()
             a.setSupportPrice(price)
         }
     }
+
+
+    fun getSupportInfo(): SupportInfo? {
+        return if (activity is MainActivity) {
+            val a = activity as MainActivity
+            if (a.supportInfo == null) null
+            else a.supportInfo
+        } else null
+    }
+
+    fun purchaseSelectItem(id: String, purchaseSuccess: (String) -> Unit, purchaseFail: () -> Unit) {
+        if (activity is MainActivity) {
+            val a = activity as MainActivity
+            a.purchaseSelectItem(id, purchaseSuccess, purchaseFail)
+        }
+    }
+
 
 }
